% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_ray_geometries}
\alias{create_ray_geometries}
\title{Create Ray Geometries for Map Visualization}
\usage{
create_ray_geometries(fetch_data)
}
\arguments{
\item{fetch_data}{Results from \code{\link{fetch_calculate}}}
}
\value{
An sf object with ray line geometries
}
\description{
Create line geometries representing fetch rays from each site.
Useful for detailed visualization of the ray-casting results.
}
\examples{
\dontrun{
results <- fetch_calculate(sites, lake)
rays <- create_ray_geometries(results)

# Plot rays for a specific site
library(ggplot2)
site_rays <- rays[rays$Site == "Site1", ]
ggplot() + geom_sf(data = site_rays, aes(color = Distance))
}

}
