% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{lakefetch-package}
\alias{lakefetch-package}
\alias{lakefetch}
\title{lakefetch: Calculate Fetch and Wave Exposure for Lake Sampling Points}
\description{
The lakefetch package provides tools for calculating fetch (open water
distance) and wave exposure metrics for lake sampling points. It downloads
lake boundaries from OpenStreetMap, calculates directional fetch using
ray-casting, and optionally integrates with NHD for hydrological context.
}
\section{Main Functions}{

\describe{
\item{\code{\link{fetch_calculate}}}{Main entry point for fetch calculation}
\item{\code{\link{load_sites}}}{Load and validate site data}
\item{\code{\link{get_lake_boundary}}}{Get lake boundary from OSM or file}
\item{\code{\link{add_lake_context}}}{Add NHD hydrological context}
\item{\code{\link{fetch_app}}}{Launch interactive Shiny app}
}
}

\section{Visualization}{

\describe{
\item{\code{\link{plot_fetch_map}}}{Map of sites colored by exposure}
\item{\code{\link{plot_fetch_bars}}}{Bar chart of effective fetch}
\item{\code{\link{plot_fetch_rose}}}{Rose diagram for single site}
\item{\code{\link{create_ray_geometries}}}{Create ray lines for mapping}
}
}

\section{Configuration}{

\describe{
\item{\code{\link{lakefetch_options}}}{Get/set package options}
\item{\code{\link{lakefetch_reset_options}}}{Reset options to defaults}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/lakefetch/}
  \item \url{https://github.com/ropensci/lakefetch}
  \item Report bugs at \url{https://github.com/ropensci/lakefetch/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Lynch Farrell \email{farrej2@rpi.edu}

}
