% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{lakefetch_reset_options}
\alias{lakefetch_reset_options}
\title{Reset lakefetch Options to Defaults}
\usage{
lakefetch_reset_options()
}
\value{
Invisible NULL
}
\description{
Reset all lakefetch package options to their default values.
}
\examples{
lakefetch_reset_options()

}
