% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_fetch_bars}
\alias{plot_fetch_bars}
\title{Plot Fetch Bar Chart}
\usage{
plot_fetch_bars(fetch_data, title = "Effective Fetch by Site")
}
\arguments{
\item{fetch_data}{Results from \code{\link{fetch_calculate}}}

\item{title}{Optional plot title}
}
\value{
A ggplot2 object
}
\description{
Create a bar chart showing effective fetch by site.
}
\examples{
\dontrun{
results <- fetch_calculate(sites, lake)
plot_fetch_bars(results)
}

}
